small animation fixes
git-svn-id: https://svn.win.tue.nl/repos/prom/Packages/InductiveVisualMiner/Trunk@21069 a68f48d4-8d00-424f-b812-3f22a534a02a

+++ src/org/processmining/plugins/inductiveVisualMiner/animation/svgToken/DotTokens2SVGtokens.java
@@ -52,14 +52,14 @@ public class DotTokens2SVGtokens {
 		svgTokens.addTrace(result.toString(), token.getLastTime(), token.isFade());
 	}
 
-	public static double fadeDuration = 0.5;
+	public static float fadeDuration = 0.5f;
 
 	public static void fade(boolean fade, boolean trueInFalseOut, Double beginTime, Double endTime, StringBuilder result) {
 		//we would use set if svgsalamander would support
 		//it does not, so we have to use an animate tag
-		double fadeDuration2;
+		float fadeDuration2;
 		if (!fade) {
-			fadeDuration2 = 0.05;
+			fadeDuration2 = 0.05f;
 			fade = true;
 		} else {
 			fadeDuration2 = fadeDuration;
